operations: support for API crawling (#20)
This PR extends support for API crawling, in particular it adds a more detailed 'Response' type to capture additional metadata about a crawled object. It also adds commonly generic utility functions such as RunCrawl. Checkpointing is also removed from the Paginator interface since a Paginator cannot be sure that its state is in sync with what the rest of the application has saved and consequently writing of checkpoints is best carried out in the 'save' portion of an application.